secret_manager_agent/requirements.txt
fastapi
uvicorn
google-cloud-secret-manager
